requirements.txt
asgiref==3.7.2
dj-database-url==2.2.0
Django==4.2.7
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.12.2
